+++ packages/core/src/world/dimension.ts
@@ -268,7 +268,7 @@ class Dimension {
             if (!item) continue;
 
             // Iterate over all the traits in the item
-            for (const [identifier, trait] of item.traits)
+            for (const trait of item.getAllTraits())
               try {
                 // Tick the item trait
                 trait.onTick?.({ currentTick, deltaTick });
@@ -279,12 +279,12 @@ class Dimension {
               } catch (reason) {
                 // Log the error to the console
                 this.world.logger.error(
-                  `Failed to tick item trait "${identifier}" for item "${item.type.identifier}" in dimension "${this.identifier}"`,
+                  `Failed to tick item trait "${trait.identifier}" for item "${item.type.identifier}" in dimension "${this.identifier}"`,
                   reason
                 );
 
                 // Remove the trait from the item
-                item.traits.delete(identifier);
+                item.removeTrait(trait.identifier);
               }
           }
         }